requirements.txt
CairoSVG==2.7.1
pillow==10.4.0
tqdm==4.66.5